refactor : LottoCalculator 클래스 및 테스트 코드 삭제, Money 클래스에 돈 관련 메서드 추가 및 테스트 코드 작성
