+++ src/test/java/domain/MoneyTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.stream.Stream;
 
 public class MoneyTest {
@@ -31,4 +33,14 @@ public class MoneyTest {
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
+    @DisplayName("수익률을 리턴해주는 메서드 테스트")
+    @ParameterizedTest
+    @CsvSource(value = {"MATCH_THREE,1,1000,500", "MATCH_FOUR,2,1000,10000", "MATCH_FIVE,3,4500,100000"})
+    void calculateProfit(RankType winningStr, int winningCount, int money, int expected) {
+        Map<RankType, Integer> winningCountMap = new HashMap<>();
+
+        winningCountMap.put(winningStr, winningCount);
+
+        Assertions.assertThat(Money.getProfit(winningCountMap, money)).isEqualTo(expected);
+    }
 }
